requirements.txt
requests==2.22.0
celery==4.4.0
redis==3.3.11
dropbox==9.4.0
boto3==1.10.39
PyYAML==5.2
python-redis-lock==3.4.0
